refactored word mapping for modularity and peformance efficiency
