+++ src/translit.js
@@ -2,6 +2,7 @@
 const languages = {};
 languages.english = require('./translit_modules/english');
 languages.english_v2 = require('./translit_modules/english_v2');
+languages.english_v3 = require('./translit_modules/english_v3');
 languages.devnagri = require('./translit_modules/devnagri');
 languages.ipa = require('./translit_modules/ipa');
 
@@ -10,6 +11,8 @@ languages.ipa = require('./translit_modules/ipa');
  *
  * @since 1.0.0
  * @param {string} gurmukhi The string from to generate transliteration
+ * @param (optional) {string} language The name of language translit module to use
+ * @param (optional) {Map} map The name of language word map to use (for per-word translit)
  * @returns {string} Returns a string of text
  * @example
  *
@@ -17,13 +20,33 @@ languages.ipa = require('./translit_modules/ipa');
  * // => 'aai mil gursikh aai mil too mere guroo ke piaare ||'
  */
 
-function translit(gurmukhi, language = 'english') {
+function translit(gurmukhi, language = 'english', map = null) {
   if (language === 'all') {
     return Object.keys(languages).reduce((out, x) => {
       /* eslint no-param-reassign: 0 */
       out[x] = languages[x](gurmukhi);
       return out;
     }, {});
+  }
+  function getMappedWord(word) {
+    if (map != null) {
+      const result = map.get(word);
+      if (result) {
+        // remove embedded vishraams before mapping
+        const vishraamRegex = /[.,;]$/;
+        const vishraam = word.match(vishraamRegex);
+        word = word.replace(vishraamRegex, '');
+        if (vishraam != null) {
+          return result + vishraam;
+        }
+        return result;
+      }
+    }
+    // use the specified translit module if word not in map
+    return languages[language](word);
+  }
+  if (map != null) {
+    return gurmukhi.split(' ').map(getMappedWord).join(' ');
   }
   return languages[language](gurmukhi);
 }
